% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\name{hs_companies_raw}
\alias{hs_companies_raw}
\alias{hs_companies_tidy}
\title{Companies endpoint (raw and tidy)}
\usage{
hs_companies_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  properties = get_company_properties(token_path, apikey),
  property_history = "true",
  max_iter = 10,
  max_properties = 100
)

hs_companies_tidy(companies = hs_companies_raw(), view = c("properties"))
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{properties}{Character vector of properties to request}

\item{property_history}{Whether version history for properties should be
returned}

\item{max_iter}{The API is limited to 250 responses, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}

\item{companies}{A companies return from \code{hs_companies_raw()}}

\item{view}{A view (see Details)}
}
\value{
List with company data (\code{hs_companies_raw()})

A tibble with associated entities (\code{hs_companies_tidy()})
}
\description{
Get raw and tidy results from \href{https://developers.hubspot.com/docs/methods/companies/get-all-companies}{the companies endpoint}.
}
\details{
Different \code{view} values and associated output.
\itemize{
\item "properties": A tibble containing all properties of companies
}
}
\examples{
\donttest{
companies <- hs_companies_raw(
  property_history = "false", max_iter = 1,
  max_properties = 10
)
companies_properties <- hs_companies_tidy(
  companies,
  view = "properties"
)
}
}
