% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies-deprecated.R
\name{company_properties}
\alias{company_properties}
\title{Extract a table of company IDs and simple property values}
\usage{
company_properties(companies = get_companies(max_iter = 1))
}
\arguments{
\item{companies}{A companies return from \code{hs_companies_raw()}}
}
\value{
A tibble with associated entities
}
\description{
Soon to be deprecated, use \code{hs_companies_tidy(companies, view = "properties")}.
}
\examples{
companies <- company_properties()
}
