% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_lists.R
\name{hs_contact_lists_raw}
\alias{hs_contact_lists_raw}
\alias{hs_contact_lists_tidy}
\title{Contact lists endpoint (raw and tidy)}
\usage{
hs_contact_lists_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  max_iter = 10
)

hs_contact_lists_tidy(
  contact_lists = hs_contact_lists_raw(),
  view = c("names")
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{max_iter}{The API is limited to a certain number of responses by page, use \code{max_iter} to
limit how many pages of data will get returned}

\item{contact_lists}{A contact lists return from \code{hs_contact_lists_raw()}}

\item{view}{A view (see Details)}
}
\value{
List of contact lists (\code{hs_contact_lists_raw()})

A tibble with associated entities (\code{hs_contact_lists_tidy()})
}
\description{
Retrieve raw and tidy results from \href{https://developers.hubspot.com/docs/methods/lists/get_lists}{the contact lists endpoint}.
}
\details{
Required scope(s) of the OAuth token: contacts.

Different \code{view} values and associated output.
\itemize{
\item "names": A tibble containing all IDs and names of contact lists
}
}
\examples{
\donttest{
contact_lists <- hs_contact_lists_raw()
hs_contact_lists_tidy(contact_lists)
}
}
