% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts-deprecated.R
\name{contact_properties}
\alias{contact_properties}
\title{Extract a table of contact IDs and simple property values}
\usage{
contact_properties(contacts = hs_contacts_raw())
}
\arguments{
\item{contacts}{A contacts return from \code{hs_contacts_raw()}}
}
\value{
A tibble with associated entities
}
\description{
Soon to be deprecated! Use \code{hs_contacts_tidy(contacts, view = "properties")}.
}
