% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{hs_contacts_raw}
\alias{hs_contacts_raw}
\alias{hs_contacts_tidy}
\title{Contacts endpoint (raw and tidy)}
\usage{
hs_contacts_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  properties = hs_contact_properties_tidy(hs_contact_properties_raw(token_path =
    token_path, apikey = apikey)),
  property_history = "true",
  form_submission_mode = "newest",
  list_memberships = "false",
  max_iter = 10,
  max_properties = 100
)

hs_contacts_tidy(contacts = hs_contacts_raw(), view = c("properties"))
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{properties}{Character vector of properties to request}

\item{property_history}{Whether version history for properties should be
returned}

\item{form_submission_mode}{Which form submissions should be fetched}

\item{list_memberships}{Indicate whether current list memberships should
be fetched}

\item{max_iter}{The API is limited to a certain number of responses by page, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}

\item{contacts}{A contacts return from \code{hs_contacts_raw()}}

\item{view}{A view (see Details)}
}
\value{
List with contact data (\code{hs_contacts_raw()})

A tibble with associated entities (\code{hs_contacts_tidy()})
}
\description{
Get raw and tidy results from the \href{https://developers.hubspot.com/docs/methods/contacts/get_contacts}{contacts endpoint}.
}
\details{
Required scope(s) of the OAuth token: contacts.

Different \code{view} values and associated output.
\itemize{
\item "properties": A tibble containing all properties of contacts
}
}
\examples{
contacts <- hs_contacts_raw(
  property_history = "false", max_iter = 1,
  max_properties = 10
)
contacts_properties <- hs_contacts_tidy(
  contacts,
  view = "properties"
)
}
