% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_pipelines.R
\name{hs_deal_pipelines_raw}
\alias{hs_deal_pipelines_raw}
\alias{hs_deal_pipelines_tidy}
\title{Deal pipelines endpoint (raw and tidy)}
\usage{
hs_deal_pipelines_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  include_inactive = "EXCLUDE_DELETED"
)

hs_deal_pipelines_tidy(
  deal_pipelines = hs_deal_pipelines_raw(),
  view = c("properties", "stages")
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{include_inactive}{Should inactive entities be returned}

\item{deal_pipelines}{A list of pipelines from \code{hs_deal_pipelines_raw}}

\item{view}{A view (see Details)}
}
\value{
List with deal pipeline data (\code{hs_deal_pipelines_raw()})

A tibble with associated entities (\code{hs_deal_pipelines_tidy()})
}
\description{
Get raw and tidy results from the \href{https://developers.hubspot.com/docs/methods/deal-pipelines/get-all-deal-pipelines}{deal pipelines endpoint}.
}
\details{
Inactive pipelines are defined as deleted pipelines.

Different \code{view} values and associated output.
\itemize{
\item "properties": tibble of pipeline IDs and property values
\item "stages": tibble of pipeline IDs and stages
}
}
\examples{
deal_pipelines <- hs_deal_pipelines_raw()
properties <- hs_deal_pipelines_tidy(
  deal_pipelines,
  view = "properties"
)
stages <- hs_deal_pipelines_tidy(
  deal_pipelines,
  view = "stages"
)
}
