% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_associations.R
\name{deal_associations}
\alias{deal_associations}
\title{Extract a table of associations from a deals list}
\usage{
deal_associations(apikey = "demo", deals = get_deals(apikey, max_iter =
  1, max_properties = 1, associations = "true"))
}
\arguments{
\item{apikey}{API key to work with Hubspot}

\item{deals}{A deals list object, typically produced by running \code{get_deals}}
}
\value{
A tibble with associated entities
}
\description{
Extract a table of associations from a deals list
}
\examples{
deal_associations()
}
