% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_pipeline_stages.R
\name{deal_pipeline_stages}
\alias{deal_pipeline_stages}
\title{Extract a table of pipeline IDs and stages}
\usage{
deal_pipeline_stages(pipelines = get_deal_pipelines())
}
\arguments{
\item{pipelines}{A list of pipelines from \code{get_deal_pipelines}}
}
\value{
A tibble of deal pipeline stages
}
\description{
Extract a table of pipeline IDs and stages
}
\examples{
deal_pipe_stages <- deal_pipeline_stages()
}
