% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_pipelines.R
\name{deal_pipelines}
\alias{deal_pipelines}
\title{Extract a table of pipeline IDs and property values}
\usage{
deal_pipelines(pipelines = get_deal_pipelines())
}
\arguments{
\item{pipelines}{A list of pipelines from \code{get_deal_pipelines}}
}
\value{
A tibble of deal pipelines
}
\description{
Extract a table of pipeline IDs and property values
}
\examples{
deal_pipes <- deal_pipelines()
}
