% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_properties.R
\name{deal_properties}
\alias{deal_properties}
\title{Extract a table of deal IDs and simple property values}
\usage{
deal_properties(deals = get_deals(max_iter = 1))
}
\arguments{
\item{deals}{A deals list object, typically produced by running \code{get_deals}}
}
\value{
A tibble with associated entities
}
\description{
Extract a table of deal IDs and simple property values
}
\examples{
deals <- deal_properties()
}
