% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deals-deprecated.R
\name{deal_properties_history}
\alias{deal_properties_history}
\title{Retrieve the changes of all properties for a given set of deals}
\usage{
deal_properties_history(
  deals = hs_deals_raw(property_history = "true", associations = "false", max_iter = 1)
)
}
\arguments{
\item{deals}{A deals list object, typically produced by running \code{hs_deals_raw}}
}
\value{
A tbl of all field changes over time
}
\description{
Soon deprecated! Use \code{hs_deals_tidy(deals, view = "properties history")}
}
