% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_stages_history.R
\name{deal_stages_history}
\alias{deal_stages_history}
\title{Retrieve the version history of deal stages}
\usage{
deal_stages_history(
  deals = get_deals(properties = "dealstage", property_history = TRUE, associations =
    "false")
)
}
\arguments{
\item{deals}{A deals list object, typically produced by running \code{get_deals}}
}
\value{
A tbl containing all stages a deal has gone through
}
\description{
Retrieve the version history of deal stages
}
\examples{
hist <- deal_stages_history()
}
