% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deals.R
\name{hs_deals_raw}
\alias{hs_deals_raw}
\alias{hs_deals_tidy}
\title{Deals endpoint (raw and tidy)}
\usage{
hs_deals_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  properties = get_deal_properties(token_path, apikey),
  property_history = "true",
  associations = "true",
  max_iter = 10,
  max_properties = 100
)

hs_deals_tidy(
  deals = hs_deals_raw(),
  view = c("associations", "properties history", "properties", "stages history")
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{properties}{Character vector of properties to request}

\item{property_history}{Whether version history for properties should be
returned}

\item{associations}{Retrieve links to other entities}

\item{max_iter}{The API is limited to 250 responses, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}

\item{deals}{A deals list object, typically produced by running \code{hs_deals_raw}}

\item{view}{A view (see Details)}
}
\value{
List with deals data (\code{hs_deals_raw()})

A tibble (\code{hs_deals_tidy()})
}
\description{
Get raw and tidy results from the \href{https://developers.hubspot.com/docs/methods/deals/get-all-deals}{deals endpoint}.
}
\details{
Different \code{view} values and associated output.
\itemize{
\item "associations": A tibble with associated entities
\item "properties history": A tibble of all field changes over time
\item "properties": A tibble containing all properties of deals
\item "stages history": A tibble containing all stages a deal has gone through
}
}
\examples{
\donttest{
deals <- hs_deals_raw(
  property_history = "false", max_iter = 1,
  max_properties = 10
)

deals_properties <- hs_deals_tidy(
  deals,
  view = "properties"
  )
}
}
