% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engagements.R
\name{hs_engagements_raw}
\alias{hs_engagements_raw}
\alias{hs_engagements_tidy}
\title{WIP - Get all engagements (raw and tidy)}
\usage{
hs_engagements_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  max_iter = 10
)

hs_engagements_tidy(engagements = hs_engagements_raw(), view = "id")
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{max_iter}{The API is limited to a certain number of responses by page, use \code{max_iter} to
limit how many pages of data will get returned}

\item{engagements}{An engagements object, typically produced by running \code{hs_engagements_raw()}}

\item{view}{A view (see Details)}
}
\value{
A list (\code{hs_engagements_raw()})

A tibble with associated entities (\code{hs_engagements_tidy()})
}
\description{
Get all of the engagements in a portal.  Returns a paginated set of engagements. from the \href{https://developers.hubspot.com/docs/methods/engagements/get-all-engagements}{Get all engagements endpoint}.
}
\details{
Required scope(s) for the OAuth token: contacts.

view = "id" returns a list of all engagements, type, created and modified dates listed by ID.

view = "contacts" also returns associated contacts, one contact ID per row

view = "companies" also returns associated companies, one company ID per row

view = "deals" also returns associated deals, one deal ID per row
}
\examples{
\donttest{
engagements <- hs_engagements_raw()
hs_engagements_tidy(engagements)
}
}
