% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-epoch_converter.R
\name{epoch_converter}
\alias{epoch_converter}
\title{Convert any potential epoch columns to datetimes}
\usage{
epoch_converter(dt)
}
\arguments{
\item{dt}{A tbl}
}
\value{
A tbl
}
\description{
Convert any potential epoch columns to datetimes
}
