% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_from_a_portal.R
\name{hs_forms_from_a_portal_raw}
\alias{hs_forms_from_a_portal_raw}
\alias{hs_froms_from_a_portal}
\alias{hs_forms_from_a_portal_tidy}
\title{WIP - Get all forms from a portal (raw and tidy)}
\usage{
hs_forms_from_a_portal_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  all_form_types = FALSE
)

hs_forms_from_a_portal_tidy(
  forms_from_a_portal = hs_forms_from_a_portal_raw(),
  view = "forms"
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{all_form_types}{By default non-marketing forms are filtered out of this endpoint. To request all forms, use this paramter with the value of TRUE.}

\item{forms_from_a_portal}{A forms_from_a_portal object, typically produced by running \code{hs_forms_from_a_portal_raw()}}

\item{view}{A view (see Details)}
}
\value{
A list (\code{hs_forms_from_a_portal_raw()})

A tibble with associated entities (\code{hs_forms_from_a_portal_tidy()})
}
\description{
For a given portal, return all forms that have been created in the portal. from the \href{https://developers.hubspot.com/docs/methods/forms/v2/get_forms}{Get all forms from a portal endpoint}.
}
\details{
Required scope(s) for the OAuth token: forms.

Different \code{view} values and associated output.
\itemize{
\item "forms": tibble of form IDs, names, createdAt dates, upatedAt date, type, published (T/F) and number of fields
\item "fields": tibble of all forms as above with form fields (one field per row)
\item "options": tibble of all forms and fields as above, with options for multiple option fields (one option per row)
}
}
\examples{
\donttest{
hs_forms_from_a_portal_raw()
forms <- hs_forms_from_a_portal_tidy()
}
}
