% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies-deprecated.R
\name{get_companies}
\alias{get_companies}
\title{Get companies}
\usage{
get_companies(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  properties = hs_company_properties_tidy(hs_company_properties_raw(token_path, apikey)),
  property_history = "true",
  max_iter = 10,
  max_properties = 100
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{properties}{Character vector of properties to request}

\item{property_history}{Whether version history for properties should be
returned}

\item{max_iter}{The API is limited to 250 responses, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}
}
\value{
List with company data
}
\description{
Soon to be deprecated! Use \code{hs_companies_raw()}.
}
\seealso{
Other getters: 
\code{\link{get_contact_properties}()},
\code{\link{get_deal_properties}()}
}
\concept{getters}
