% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_properties.R
\name{get_company_properties}
\alias{get_company_properties}
\title{Retrieve available properties for companies}
\usage{
get_company_properties(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL})}
}
\value{
Character vector of properties
}
\description{
Retrieve available properties for companies
}
\examples{
properties <- get_company_properties()
}
\seealso{
Other getters: 
\code{\link{get_companies}()},
\code{\link{get_contact_properties}()},
\code{\link{get_contacts}()},
\code{\link{get_deal_pipelines}()},
\code{\link{get_deal_properties}()},
\code{\link{get_deals}()},
\code{\link{get_owners}()}
}
\concept{getters}
