% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts_properties.R
\name{get_contact_properties}
\alias{get_contact_properties}
\title{Retrieve available properties for contacts}
\usage{
get_contact_properties(apikey = hubspot_key_get())
}
\arguments{
\item{apikey}{API key to work with Hubspot}
}
\value{
Character vector of properties
}
\description{
Retrieve available properties for contacts
}
\examples{
properties <- get_contact_properties()
}
\seealso{
Other getters: \code{\link{get_companies}},
  \code{\link{get_company_properties}},
  \code{\link{get_contacts}},
  \code{\link{get_deal_pipelines}},
  \code{\link{get_deal_properties}},
  \code{\link{get_deals}}, \code{\link{get_owners}}
}
\concept{getters}
