% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts-deprecated.R
\name{get_contacts}
\alias{get_contacts}
\title{Get a list of all contacts, including a set of properties}
\usage{
get_contacts(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  properties = get_contact_properties(token_path = token_path, apikey = apikey),
  property_history = "true",
  form_submission_mode = "newest",
  list_memberships = "false",
  max_iter = 10,
  max_properties = 100
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{properties}{Character vector of properties to request}

\item{property_history}{Whether version history for properties should be
returned}

\item{form_submission_mode}{Which form submissions should be fetched}

\item{list_memberships}{Indicate whether current list memberships should
be fetched}

\item{max_iter}{The API is limited to 250 responses, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}
}
\value{
List with contact data
}
\description{
Soon to be deprecated! Use \code{hs_contacts_raw()}.
}
