% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_pipelines-deprecated.R
\name{get_deal_pipelines}
\alias{get_deal_pipelines}
\title{Retrieve available pipelines for deals}
\usage{
get_deal_pipelines(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  include_inactive = "EXCLUDE_DELETED"
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{include_inactive}{Should inactive entities be returned}
}
\value{
List with deal pipeline data
}
\description{
Soon to be deprecated! Use \code{hs_deal_pipelines_raw()}.
}
\details{
Inactive pipelines are defined as deleted pipelines.
}
