% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_owners.R
\name{get_owners}
\alias{get_owners}
\title{Retrieve owners}
\usage{
get_owners(apikey = hubspot_key_get())
}
\arguments{
\item{apikey}{API key to work with Hubspot}
}
\value{
List of owners
}
\description{
Retrieve owners
}
\examples{
owners <- get_owners()
}
\seealso{
Other getters: \code{\link{get_companies}},
  \code{\link{get_company_properties}},
  \code{\link{get_contact_properties}},
  \code{\link{get_contacts}},
  \code{\link{get_deal_pipelines}},
  \code{\link{get_deal_properties}},
  \code{\link{get_deals}}
}
\concept{getters}
