% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_properties.R
\name{hs_company_properties_raw}
\alias{hs_company_properties_raw}
\title{Retrieve available properties for companies}
\usage{
hs_company_properties_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
Character vector of properties
}
\description{
Retrieve available properties for companies from \href{https://developers.hubspot.com/docs/methods/companies/}{the company properties endpoint}.
}
\examples{
properties <- hs_company_properties_raw()
}
