% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{hubspot-deprecated}
\alias{hubspot-deprecated}
\title{Deprecated functions in hubspot}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item \code{get_deals}
\item \code{deal_properties_history}
}
}
