% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot_key.R
\name{hubspot_key_get}
\alias{hubspot_key_get}
\alias{hubspot_key_set}
\title{Get/set the Hubspot secret API key}
\usage{
hubspot_key_get()

hubspot_key_set(key = NULL)
}
\arguments{
\item{key}{API key. If not provided via this argument, the key will have
to be provided interactively.}
}
\value{
The key (a string)
}
\description{
Get/set the Hubspot secret API key using the \code{keyring} package.
For a more secure authorization method, without a daily limit on calls,
see \code{\link[=hubspot_token_get]{hubspot_token_get()}}
}
\details{
To know how to get your Hubspot API key, refer to \href{https://knowledge.hubspot.com/integrations/how-do-i-get-my-hubspot-api-key}{the corresponding section of Hubspot API docs}.
Hubspot recommends you to re-generate your API key every six months.

The API key should be set using the \code{hubspot_key_set()} function, that
relies on the \code{keyring} package that stores the key in the system
default credential store (see \code{keyring} docs).\if{html}{\out{<div class="r">}}\preformatted{# Either input the key as parameter
hubspot_key_set("yourapikey")

# Or just call the function,
# the key will be asked interactively
hubspot_key_set()
}\if{html}{\out{</div>}}

If no API key is set, the “demo” key is used and a message is printed to
the screen. As noted earlier, this won’t give you access to your own
Hubspot data.

You can check the daily usage related to the API key via
\code{hubspot_api_usage()}. Calls in the package are rate limited (100
requests every 10 seconds) but there’s also a daily limit on calls.\if{html}{\out{<div class="r">}}\preformatted{hubspot_api_usage()
}\if{html}{\out{</div>}}

\strong{For a more secure authorization method, without a daily limit on
calls, see \code{\link[=hubspot_token_get]{hubspot_token_get()}}.}

\subsection{Authorization for Hubspot API}{

The Hubspot API accepts authorization via
\itemize{
\item an API key,
\item OAuth 2.0.
}

OAuth 2.0 is the \href{https://developers.hubspot.com/docs/methods/auth/oauth-overview?_ga=2.108539650.1064389456.1574673541-1134397846.1571640267}{recommended method}.
However, this package supports both.

Note that if you do nothing the package will use the “demo” API token
but this won’t give you access to your own Hubspot data. So you’ll need
to spend a little time on setup:
\itemize{
\item For rapid prototyping key, use a Hubspot API key, see
\code{hubspot_key_set()}.
\item For more secure use, without a daily limit on API calls, see
\code{hubspot_token_get()} to create a Hubspot authorization token (OAuth
2.0).
}

If you have both saved an API key via \code{hubspot_key_set()} and a token
via \code{hubspot_token_get()}, priority will be given to using the OAuth 2.0
token. If you don’t want that, explicitely pass BLABLA
}
}
\examples{
hubspot_key_get()
\dontrun{
hubspot_key_set("yourapikey")
}
}
\seealso{
Other auth: 
\code{\link{hubspot_token_get}()}
}
\concept{auth}
