% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_usage.R
\name{hubspot_api_usage}
\alias{hubspot_api_usage}
\title{Check daily API usage}
\usage{
hubspot_api_usage(apikey = hubspot_key_get())
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}
}
\value{
A data.frame with API usage info
}
\description{
Check daily API usage \href{https://developers.hubspot.com/docs/methods/check-daily-api-usage}{via the dedicated API endpoint}.
}
\examples{
\dontrun{
hubspot_api_usage()
}
}
