% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot_key.R
\name{hubspot_key_set}
\alias{hubspot_key_set}
\title{Set the Hubspot secret API key}
\usage{
hubspot_key_set(key = NULL)
}
\arguments{
\item{key}{API key. If not provided via this argument, the key will have
to be provided interactively.}
}
\description{
Set the Hubspot secret API key using the \code{keyring} package.
}
\examples{
\dontrun{
hubspot_key_set("yourapikey")
}
}
\seealso{
Other auth: 
\code{\link{hubspot_key_get}()}
}
\concept{auth}
