% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot_oauth.R
\name{hubspot_token_get}
\alias{hubspot_token_get}
\title{Create Hubspot authorization token.}
\usage{
hubspot_token_get(app_info = ld_hubspot_app())
}
\arguments{
\item{app_info}{A named list with client_secret, client_id, app_id and
scopes.}
}
\value{

}
\description{
Create Hubspot authorization token.
}
\details{
We expect you to not \emph{need} to bring your own app (BYOA), but the
\code{app_info} parameter of \code{hubspot_token_get()} ensures you can do that if
you want (for instance to define an app with only the scope needed for
the endpoints you use as opposed to all the endpoints the \code{hubspot}
package supports).

To BYOA,
\itemize{
\item \href{https://developers.hubspot.com/docs/faq/how-do-i-create-an-app-in-hubspot}{Read and execute Hubspot’s docs about how to create an app}.
\item Define the \code{app_info} list.
}\if{html}{\out{
<!-- end list -->
}}
\if{html}{\out{<div class="r">}}\preformatted{my_very_own_app <- list(client_secret = <CLIENTSECRET>,
       client_id = <CLIENTID>,
       app_id = <APPID>,
       scope = c(<"scope1", "scope2", etc>))
}\if{html}{\out{</div>}}
\itemize{
\item Pass the app info to \code{hubspot_token_get()}
}\if{html}{\out{
<!-- end list -->
}}
\if{html}{\out{<div class="r">}}\preformatted{hubspot_token_get(app_info = my_very_own_app)
}\if{html}{\out{</div>}}

To distribute your app across your projects or across team members, you
could make it a function in an internal package, or use environment
variables.

\subsection{Authorization for Hubspot API}{

The Hubspot API accepts authorization via
\itemize{
\item an API key,
\item OAuth 2.0.
}

OAuth 2.0 is the \href{https://developers.hubspot.com/docs/methods/auth/oauth-overview?_ga=2.108539650.1064389456.1574673541-1134397846.1571640267}{recommended method}.
However, this package supports both.

Note that if you do nothing the package will use the “demo” API token
but this won’t give you access to your own Hubspot data. So you’ll need
to spend a little time on setup:
\itemize{
\item For rapid prototyping key, use a Hubspot API key, see
\code{hubspot_key_set()}.
\item For more secure use, without a daily limit on API calls, see
\code{hubspot_token_get()} to create a Hubspot authorization token (OAuth
2.0).
}

If you have both saved an API key via \code{hubspot_key_set()} and a token
via \code{hubspot_token_get()}, priority will be given to using the OAuth 2.0
token. If you don’t want that, explicitely pass BLABLA
}
}
\seealso{
Other auth: 
\code{\link{hubspot_key_get}()}
}
\concept{auth}
