% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_properties.R
\name{company_properties}
\alias{company_properties}
\title{Extract a table of company IDs and simple property values}
\usage{
company_properties(companies = get_companies(max_iter = 1))
}
\arguments{
\item{companies}{A companies return from \code{get_companies()}}
}
\value{
A tibble with associated entities
}
\description{
Extract a table of company IDs and simple property values
}
\examples{
companies <- company_properties()
}
