% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_properties.R
\name{contact_properties}
\alias{contact_properties}
\title{Extract a table of contact IDs and simple property values}
\usage{
contact_properties(contacts = get_contacts(max_iter = 1))
}
\arguments{
\item{contacts}{A contacts return from \code{get_contacts()}}
}
\value{
A tibble with associated entities
}
\description{
Extract a table of contact IDs and simple property values
}
\examples{
contacts <- contact_properties()
}
