% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deals-deprecated.R
\name{deal_associations}
\alias{deal_associations}
\title{Extract a table of associations from a deals list}
\usage{
deal_associations(
  deals = get_deals(max_iter = 10, max_properties = 1, associations = "true")
)
}
\arguments{
\item{deals}{A deals list object, typically produced by running \code{hs_deals_raw}}
}
\value{
A tibble with associated entities
}
\description{
Soon deprecated! Use \code{hs_deals_tidy(deals, view = "associations")}
}
\examples{
assoc <- deal_associations()
}
