% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_associations.R
\name{deal_associations}
\alias{deal_associations}
\title{Extract a table of associations from a deals list}
\usage{
deal_associations(deals = get_deals(max_iter = 10, max_properties = 1,
  associations = "true"))
}
\arguments{
\item{deals}{A deals list object, typically produced by running \code{get_deals}}
}
\value{
A tibble with associated entities
}
\description{
Extract a table of associations from a deals list
}
\examples{
assoc <- deal_associations()
}
