% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_pipelines-deprecated.R
\name{deal_pipeline_stages}
\alias{deal_pipeline_stages}
\title{Extract a table of pipeline IDs and stages}
\usage{
deal_pipeline_stages(deal_pipelines = hs_deal_pipelines_raw())
}
\arguments{
\item{deal_pipelines}{A list of pipelines from \code{hs_deal_pipelines_raw}}
}
\value{
A tibble of deal pipeline stages
}
\description{
Soon to be deprecated! Use \code{hs_deal_pipelines_tidy(pipelines, view = "stages")}.
}
