% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_pipelines-deprecated.R
\name{deal_pipelines}
\alias{deal_pipelines}
\title{Extract a table of pipeline IDs and property values}
\usage{
deal_pipelines(deal_pipelines = hs_deal_pipelines_raw())
}
\arguments{
\item{deal_pipelines}{A list of pipelines from \code{hs_deal_pipelines_raw}}
}
\value{
A tibble of deal pipelines
}
\description{
Soon to be deprecated! Use \code{hs_deal_pipelines_tidy(pipelines, view = "properties")}.
}
