% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deals-deprecated.R
\name{deal_properties}
\alias{deal_properties}
\title{Extract a table of deal IDs and simple property values}
\usage{
deal_properties(deals = get_deals(max_iter = 1))
}
\arguments{
\item{deals}{A deals list object, typically produced by running \code{hs_deals_raw}}
}
\value{
A tibble with associated entities
}
\description{
Soon deprecated! Use \code{hs_deals_tidy(deals, view = "properties")}
}
\examples{
deals <- deal_properties()
}
