% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_properties_history.R
\name{deal_properties_history}
\alias{deal_properties_history}
\title{Retrieve the changes of all properties for a given set of deals}
\usage{
deal_properties_history(deals = get_deals(property_history = "true",
  associations = "false", max_iter = 1))
}
\arguments{
\item{deals}{A deals list object, typically produced by running \code{get_deals}}
}
\value{
A tbl of all field changes over time
}
\description{
Retrieve the changes of all properties for a given set of deals
}
\examples{
hist <- deal_properties_history()
}
