% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_properties-deprecated.R
\name{get_company_properties}
\alias{get_company_properties}
\title{Retrieve available properties for companies}
\usage{
get_company_properties(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
Character vector of properties
}
\description{
Soon to be deprecated! Use \code{hs_company_properties_tidy()}.
}
\examples{
properties <- hs_company_properties_raw()
}
