% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_properties.R
\name{get_company_properties}
\alias{get_company_properties}
\title{Retrieve available properties for companies}
\usage{
get_company_properties(apikey = "demo")
}
\arguments{
\item{apikey}{API key to work with Hubspot}
}
\value{
Character vector of properties
}
\description{
Retrieve available properties for companies
}
\examples{
properties <- get_company_properties()
}
\seealso{
Other getters: \code{\link{get_companies}},
  \code{\link{get_deal_properties}},
  \code{\link{get_deals}}
}
\concept{getters}
