% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts_properties.R
\name{get_contact_properties}
\alias{get_contact_properties}
\title{Retrieve available properties for contacts}
\usage{
get_contact_properties(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
Character vector of properties
}
\description{
Retrieve available properties for contacts
}
\examples{
properties <- get_contact_properties()
}
\seealso{
Other getters: 
\code{\link{get_companies}()},
\code{\link{get_deal_properties}()},
\code{\link{get_owners}()}
}
\concept{getters}
