% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts.R
\name{get_contacts}
\alias{get_contacts}
\title{Get a list of all contacts, including a set of properties}
\usage{
get_contacts(apikey = hubspot_key_get(),
  properties = get_contact_properties(apikey),
  property_history = "true", form_submission_mode = "newest",
  list_memberships = "false", max_iter = 10, max_properties = 100)
}
\arguments{
\item{apikey}{API key to work with Hubspot}

\item{properties}{Character vector of properties to request}

\item{property_history}{Whether version history for properties should be
returned}

\item{form_submission_mode}{Which form submissions should be fetched}

\item{list_memberships}{Indicate whether current list memberships should
be fetched}

\item{max_iter}{The API is limited to 250 responses, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}
}
\value{
List with contact data
}
\description{
Get a list of all contacts, including a set of properties
}
\examples{
contacts <- get_contacts(property_history = "false", max_iter = 1,
                         max_properties = 10)
}
\seealso{
Other getters: \code{\link{get_companies}},
  \code{\link{get_company_properties}},
  \code{\link{get_contact_properties}},
  \code{\link{get_deal_pipelines}},
  \code{\link{get_deal_properties}},
  \code{\link{get_deals}}, \code{\link{get_owners}}
}
\concept{getters}
