% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deals.R
\name{get_deals}
\alias{get_deals}
\title{Get a list of all deals, including a set of properties}
\usage{
get_deals(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  properties = get_deal_properties(token_path, apikey),
  property_history = TRUE,
  associations = "true",
  max_iter = 10,
  max_properties = 100
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{properties}{Character vector of properties to request}

\item{property_history}{Whether version history for properties should be
returned}

\item{associations}{Retrieve links to other entities}

\item{max_iter}{The API is limited to 250 responses, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}
}
\value{
List with deals data
}
\description{
Get a list of all deals, including a set of properties
}
\examples{
deals <- get_deals(
  property_history = FALSE, max_iter = 1,
  max_properties = 10
)
}
\seealso{
Other getters: 
\code{\link{get_companies}()},
\code{\link{get_company_properties}()},
\code{\link{get_contact_properties}()},
\code{\link{get_contacts}()},
\code{\link{get_deal_pipelines}()},
\code{\link{get_deal_properties}()},
\code{\link{get_owners}()}
}
\concept{getters}
