% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_owners.R
\name{get_owners}
\alias{get_owners}
\title{Retrieve owners}
\usage{
get_owners(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  include_inactive = TRUE,
  email = NULL
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL})}

\item{include_inactive}{Boolean indicating whether to include inactive
owners (defined as owners without any active remotes.)}

\item{email}{Search for owners matching the specified email address
(exact match).}
}
\value{
List of owners
}
\description{
Retrieve owners \href{https://developers.hubspot.com/docs/methods/owners/get_owners}{from the owners API endpoint}.
}
\examples{
\dontrun{
owners <- get_owners()
active_owners <- get_owners(include_inactive = FALSE)
get_owners(email = "demo@hubspot.com")
}
}
\seealso{
Other getters: 
\code{\link{get_companies}()},
\code{\link{get_company_properties}()},
\code{\link{get_contact_properties}()},
\code{\link{get_contacts}()},
\code{\link{get_deal_pipelines}()},
\code{\link{get_deal_properties}()},
\code{\link{get_deals}()}
}
\concept{getters}
