% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{hubspot-deprecated}
\alias{hubspot-deprecated}
\title{Deprecated functions in hubspot}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item \code{get_deals()}
\item \code{deal_properties_history()}
\item \code{deal_properties()}
\item \code{deal_stages_history()}
\item \code{deal_associations()}
\item \code{get_companies()}
\item \code{company_properties()}
\item \code{get_company_properties()}
\item \code{get_contacts()}
\item \code{contact_properties()}
\item \code{get_deal_pipelines()}
\item \code{deal_pipeline_stages()}
\item \code{deal_pipelines()}
\item \code{get_owners()}
\item \code{owners()}
\item \code{get_contact_properties()}
\item \code{get_deal_properties()}
}
}
