% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot-package.R
\docType{package}
\name{hubspot-package}
\alias{hubspot}
\alias{hubspot-package}
\title{hubspot: Work with Hubspot data}
\description{
Utilising the 'Hubspot CRM' API (see <https://developers.hubspot.com/docs>
    for more information), retrieve customer relationship data for reporting and analysis.
}
\author{
\strong{Maintainer}: Steph Locke \email{steph@itsalocke.com}

}
\keyword{internal}
