% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_usage.R
\name{hubspot_api_usage}
\alias{hubspot_api_usage}
\title{Check daily API usage}
\usage{
hubspot_api_usage(apikey = hubspot_key_get())
}
\arguments{
\item{apikey}{API key to work with Hubspot}
}
\value{
A data.frame with API usage info
}
\description{
Check daily API usage \href{https://developers.hubspot.com/docs/methods/check-daily-api-usage}{via the dedicated API endpoint}.
}
\examples{
\dontrun{
hubspot_api_usage()
}
}
