% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot_auth.R
\name{hubspot_token_get}
\alias{hubspot_token_get}
\title{Create Hubspot authorization token.}
\usage{
hubspot_token_get()
}
\value{

}
\description{
Create Hubspot authorization token.
}
\details{
\subsection{Authorization for Hubspot API}{

The Hubspot API accepts authorization via
\itemize{
\item an API key,
\item OAuth 2.0.
}

OAuth 2.0 is the \href{https://developers.hubspot.com/docs/methods/auth/oauth-overview?_ga=2.108539650.1064389456.1574673541-1134397846.1571640267}{recommended method}.
However, this package supports both.

Note that if you do nothing the package will use the “demo” API token
but this won’t give you access to your own Hubspot data. So you’ll need
to spend a little time on setup:
\itemize{
\item For rapid prototyping key, use a Hubspot API key, see
\code{hubspot_key_set()}.
\item For more secure use, without a daily limit on API calls, see
\code{hubspot_token_get()} to create a Hubspot authorization token (OAuth
2.0).
}

If you have both saved an API key via \code{hubspot_key_set()} and a token
via \code{hubspot_token_get()}, priority will be given to using the OAuth 2.0
token. If you don’t want that, explicitely pass BLABLA
}
}
\seealso{
Other auth: 
\code{\link{hubspot_key_get}()}
}
\concept{auth}
