% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owners.R
\name{owners}
\alias{owners}
\title{Extract a table of CRM owners}
\usage{
owners(owners = get_owners())
}
\arguments{
\item{owners}{Raw data from \code{get_owners()}}
}
\value{
data.frame of owners
}
\description{
Extract a table of CRM owners
}
\examples{
crm_owners <- owners()
}
