% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owners-deprecated.R, R/owners.R
\name{owners}
\alias{owners}
\alias{hs_owners_tidy}
\title{Extract a table of CRM owners}
\usage{
owners(owners = hs_owners_raw())

hs_owners_tidy(owners = hs_owners_raw(), view = c("owners"))
}
\arguments{
\item{owners}{Raw data from \code{hs_owners_raw()}}

\item{view}{A view (see Details)}
}
\value{
tibble of owners

A tibble of owners (\code{hs_owners_tidy()})
}
\description{
Soon to be deprecated! Use \code{hs_owners_tidy(owners, view = "owners")}.
}
\details{
Different \code{view} values and associated output.
\itemize{
\item "owners": A tibble with owners and related information.
}
}
